requirements.txt
pillow
requests
requests[socks]
pysocks